refactor: move functions out of content index.ts files, move tests under /content
